Wavesurfer init + play button rendering put into new file
Added wavesurferMethods, a class with all the methods ever needed to render the wavesurfer code and buttons.

Currently debating about adding a class to handle labels
